Restructured i18n classes
